Working changes on the renderer. Metal renderer should be close to testing grade. Added sysfunction for entry point management, such as for iOS, OSX, Android etc...
Rendering is divided into a hierarchy as such:

PipelineState encapsulates the settings related to rendering inputs and outputs, which is shaders, render target and blend states.
RenderState encapsulates the stencil and depth settings as well as rasterisation stuff.
Primitive is just a draw call descriptor, which describes which sub mesh we should render from the vertex buffer/index buffer.
And that's it, setup is super simple.

RenderStates are static and can be updated through the Commit call. This is because render states should be treated as resources, and if needed, one should copy a render state, mod it and use the copy instead of modifying the original. Does not have to be initialised prior to usage, but will instead just use its default settings.
PipelineStates has to be initialized to use either a render type pipeline or a compute one. They also have to be fed with shaders, a render target, and blend states. If the blend states and render target attachments mismatch, the pipeline state will just create empty blend states (default settings) and use them instead. As such, BlendState provides a constructor for use with nullptr, for convenience.

There is still no flexible way to initialise a vertex buffer. Considering just making the buffer init functions public, since buffers no longer have any intrinsic usages in new rendering APIs...